fix: 해커톤 병합 누락된 부분 수정
