+++ src/main/common/exception/CodeInterface.java
@@ -1,0 +1,6 @@
+package com.debateseason_backend_v1.common.exception;
+
+public interface CodeInterface {
+    Integer getCode();
+    String getMessage();
+}
