--- src/main/java/com/debateseason_backend_v1/common/exception/CodeInterface.java
@@ -1,6 +1,0 @@
-package com.debateseason_backend_v1.common.exception;
-
-public interface CodeInterface {
-    Integer getCode();
-    String getMessage();
-}
