+++ src/main/resources/application-secret.yml
@@ -1,0 +1,6 @@
+spring:
+  datasource:
+    driver-class-name: org.mariadb.jdbc.Driver
+    url: jdbc:mariadb://localhost:3306/test_database?characterEncoding=UTF-8&serverTimezone=Asia/Seoul
+    username: tester
+    password: "tester1234"
